+++ app/Services/Search/Drivers/ManticoreSearchDriver.php
@@ -847,10 +847,10 @@ class ManticoreSearchDriver implements SearchDriverInterface
                             'adddate_ts' => ['type' => 'bigint'],
                             'totalpart' => ['type' => 'integer'],
                             'grabs' => ['type' => 'integer'],
-                            'passwordstatus' => ['type' => 'integer'],
+                            'passwordstatus' => ['type' => 'bigint'],
                             'groups_id' => ['type' => 'integer'],
                             'nzbstatus' => ['type' => 'integer'],
-                            'haspreview' => ['type' => 'integer'],
+                            'haspreview' => ['type' => 'bigint'],
                         ],
                     ],
                 ]);
